refactor: decompose ArmyValidator into focused validator modules (#226)
